added Match class and Finite State Machine
Match class runs basic game rule functionality, including checking for
wins; It's difficult to test right now given that as of this moment we
can't seem to move pieces, but once that can be done, the game should be
playable.

Sorry to break the golden rule, but I had to add a very small number of
lines of code to Margaret's class files Player, Unit, and UnitClasses in
order to get things up and running.

Made the Texture2D's containing Nick's artwork public static so I could
use them in my Match class and not create all of our units in Game1.

Really simple Finite State Machine now exists in Game1's Update method.

Bugs that need to get patched up:
- For some reason, clicking on the game board after waiting a certain
amount of time seems to crash the game.
- Once at the win screen, clicking Return to Main Menu brings me right
back to the win screen after flashing the board. I'm not quite sure if
this one is my fault or Nick's, but I'm about 99.9% certain it's mine.